requirements.txt
kivy>=2.3.0
kivymd>=1.2.0
clear>=1.2.0
buildozer>=1.5.0
Cython>=3.0.10
virtualenv>=20.26.2
python-for-android==2024.1.21
